Check if product has linked items when deleting (if so, error "product.notEmptyWhenDeleting")
git-svn-id: http://agilefant.googlecode.com/svn/trunk@424 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ src/fi/hut/soberit/agilefant/web/ProductAction.java
@@ -24,6 +24,11 @@ public class ProductAction extends ActionSupport implements CRUDAction {
 	}
 
 	public String delete() {
+		Product p = productDAO.get(productId);
+		if(p.getBacklogItems().size() > 0 || p.getDeliverables().size() > 0) {
+			super.addActionError(super.getText("product.notEmptyWhenDeleting"));
+			return Action.ERROR;
+		}
 		productDAO.remove(productId);
 		return Action.SUCCESS;
 	}
